+++ src/com/xxdb/data/BasicDecimal64.java
@@ -52,7 +52,7 @@ public class BasicDecimal64 extends AbstractScalar implements Comparable<BasicDe
                     sb.append("0");
                 }
             }else {
-                sb.append(data);
+                sb.append((int)data);
             }
             return sb.toString();
         }
